Moved to a more stable RTL8812AU driver repo...
Most of the repos are derived from the vendor upstream code cleaned up
up to either support versions not supported by Realtek or for DKMS
and the requisite cleanups.  Some of these don't play nice on ARM
so I went through and made SURE the one in my metadata layer at least
worked with the 4.9 kernel branch Raspbian and The Pi Foundation
are using with the Raspberry Pi for Pi2 or Pi3 in 32-bits.  This is
kind of the latest version of the Realtek drivers cleaned up and it
seems to play nicely on Pi2 and Pi3.