Added aircrack-ng version of RTL8812AU support
This adds monitor and injection support.  5.1.5 version, per what they
default to over on Github.  5.2.9's not "released" yet and master's the
wrecking ball branch (like it SHOULD be...), so we'll use that.

This conflicts with the rtl8812au recipe and you should build one or
the other, and make sure you're cleaned out before building the other
for whatever reason as Yocto/Poky will bitch at you if you don't.