General cleanups to make this easier to maintain.
We had a raftload (and still do) of metadata that was repeated all over
the place in this recipe set- so I folded the repetitive pieces into
a single include (realtek.inc) that holds the bulk of the magic.

Used it to fix the rtl8812/8814/8822 recipes.  Will review the others
shortly- but in this case, I actively USE the others, so they get
the treatment FIRST...  X-D