Generic cleanup.
Trying to find the best path for a clean build on all of these.  While
initially forked from someone on GitHub (and this is hosted on the same)
it's become a beastie all of it's very own.  Some won't build with the
kernel provided for the NanoPI Neo2/Neo-Plus2 board I'm currently
focused on for my project work, so a review to see if we're in mainline
with something like it, can point to a cleaned up version like I've
done with the 8812/8814/8822 series parts, fork the source if neither
of the above fits.  If kernel mainline has a good answer (i.e.
RTL9182EU is there but this doesn't bring an ability to properly
provide dual-roles on the part...) then we nuke the recipe.