Adjustments to mesh with Rocko/Sumo build environment.
Proper fix was to specify a modules destination dir, override the target
for install on make since Realtek's got a definitive set of targets
there that doesn't conform, and then stage in as prepend the directory
structure expected because it is non-standard.  Seems to work all the
time.  We'll see later because this needs to go in for a validation
run in other build configs with this metadata in place without hand
holding.