Patching makefile slightly...
While we overrode the install in the previous commit, we're explicitly
turning off something in select makefiles.  We're doing that this way
so we're not making a half dozen or more fixes to broken makes- these
are targeted and the framework for OE is intended to be as generic for
all of the Realtek drivers as is possible.