Explicitly enforce the dependency to be LATE in the build process.
There are some kernels that don't set up properly for out-of-tree builds
in OE.  The meta-sunxi kernel recipe is an example of this.  It's being
compensated for as a post modules build so the right pieces of info and
config are in the right places in the shared workdir space like it's
SUPPOSED to be- but this means we need to defer OUR build of the drivers
here to being really later in the virtual/kernel build process to ensure
we start at the right time for this.